requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
mysqlclient==2.2.4
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==2.8.2
email-validator==2.1.1
PyMySQL==1.1.1
python-dotenv
pymysql
dotenv
pydantic
sqlalchemy
bcrypt==4.1.2
python-multipart
reportlab>=4.1.0 
tzdata>=2024.1
weasyprint>=60
python-barcode
pillow
jinja2
qrcode
openpyxl